+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-1-22-19
+3-22-19
 
     - Created Changelog
 
@@ -11,3 +11,22 @@
      of 800px and with a margin of 0 auto. 
 
     - Made contact form.
+
+3-25-19
+
+    - Created git branch for Contact form
+
+    - Added initial CSS for Contact form.
+
+    - Added Viewport Meta to header.php
+
+4-4-19
+
+    -Created incl and js folders
+
+    -Created incl/contact.incl.php
+
+    -Created inc/dbh.incl.php
+    
+    -Added action to contact form to point it to incl/contact.incl.php and method = to post
+